requirements.txt
Django==1.10
djangorestframework==3.5.3
factory-boy==2.8.1
ipython==5.1.0
isort==4.2.5
psycopg2==2.6.2
Sphinx==1.4.8
shortuuid==0.5.0
django-shortuuidfield==0.1.3
django-extensions==1.7.8
